requirements.txt
django==4.2.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-taggit==5.0.1
pillow==10.2.0
djangorestframework==3.15.1
djoser==2.2.2
djangorestframework_simplejwt==5.3.1
django-filter==23.5
dynamic-rest==2.3.0
drf-yasg==1.21.7
requests==2.31
django-ordered-model==3.7.4
celery[redis]==5.3.6
weasyprint==61.2
django-cors-headers==4.3.1
aiogram==3.4.1
Scrapy==2.11.1
beautifulsoup4==4.12.3
fake-headers==1.0.2
